Initial database stuff, Models folder updated
Added models for the database, and a DbContext for creating the database. No real database usage supported yet, work in progress, this is just an outline. Update-database might have to be entered to the package manager console, as well as changing the DbContext file for the appropriate MySQL username and password